refactor: create useScenario custom hook
src/pages/Start.jsx에서 기존에 하드 코딩 되어있던 scenarioPhaseState와 관련 함수들을 하나로 묶어 useScenario custom hook으로 재정의.
