+++ src/pages/Start/Start.jsx
@@ -1,7 +1,8 @@
-import React, { useState } from 'react';
+import React from 'react';
 
 import scenario from '@/data/scenario.json';
 import useConfig from '@/hooks/useConfig';
+import useScenario from '@/hooks/useScenario';
 
 import FooterL from './FooterL';
 import FooterM from './FooterM';
@@ -18,46 +19,8 @@ import './Start.scss';
  */
 function Start() {
   /* Hooks */
-  // useConfig
   const config = useConfig();
-  // scenarioPhase
-  const [scenarioPhaseState, setScenarioPhaseState] = useState({
-    major: 0,
-    minor: 0,
-  });
-  const handleNextScenarioPhaseState = type => {
-    const newScenarioPhaseState = scenarioPhaseState[type] + 1;
-
-    if (type === 'major' && newScenarioPhaseState < scenario.phase.length) {
-      setScenarioPhaseState(prevState => ({ ...prevState, major: newScenarioPhaseState, minor: 0 }));
-    } else if (type === 'minor' && newScenarioPhaseState < scenario.phase[scenarioPhaseState.major].length) {
-      setScenarioPhaseState(prevState => ({ ...prevState, [type]: newScenarioPhaseState }));
-    }
-  };
-  const handleSkipScenarioPhaseState = () => {
-    // for minor phase
-    const newScenarioPhaseState = scenario.phase[scenarioPhaseState.major].length - 1;
-
-    setScenarioPhaseState(prevState => ({ ...prevState, minor: newScenarioPhaseState }));
-  };
-  const isScenarioPhaseDone = () => {
-    // for minor phase
-    return scenarioPhaseState.minor === scenario.phase[scenarioPhaseState.major].length - 1;
-  };
-  const isTutorialPhase = () => {
-    return scenarioPhaseState.major === 0;
-  };
-  const isInterviewPhase = () => {
-    return scenarioPhaseState.major === 1;
-  };
-  const scenarioPhase = {
-    scenarioPhaseState,
-    handleNextScenarioPhaseState,
-    handleSkipScenarioPhaseState,
-    isScenarioPhaseDone,
-    isTutorialPhase,
-    isInterviewPhase,
-  };
+  const scenarioPhase = useScenario();
 
   // useQuestionMain
   // const [arrayCsQuestionMain, generateCsQuestionMain] = useQuestionMain('cs', 3);
